api::abi::syscall

Module cap_mask

Bit constants for the cap_mask argument of ViSyscall::CapRevoke.

The 32-bit mask encodes which capability classes to strip from the target cell. Boolean caps occupy the low 4 bits; parameterized caps use two 8-bit sub-fields.

Constants

ALL

Revoke all capability classes at once.

BLKREGION_MASK

Mask selecting the block_regions 8-bit sub-field (bits 16-23).

BLKREGION_SHIFT

Bit offset of the block_regions sub-field within the mask.

BLOCK_IO

Revoke raw block-device access (BlkRead/BlkWrite/BlkFlush).

HYPERVISOR

Revoke RISC-V H-extension / ARM64 EL2 hypervisor access.

MMIO_MASK

Mask selecting the mmio_devices 8-bit sub-field (bits 8-15). Each bit within this byte corresponds to a resource_registry::DEV_* flag.

MMIO_SHIFT

Bit offset of the mmio_devices sub-field within the mask.

NETWORK

Revoke network transmit/receive (NetTx/NetRx).

SPAWN

Revoke cell-spawning authority (SpawnFromPath/HotSwap).