Module api::db

Contiene estructuras y funciones para la persistencia de los datos de los clientes y sus balances en archivos y en memoria en una base de datos. Implementa el patrón de diseño Abstract Factory para poder reemplazar los métodos que pueden llevarse a cabo con la base de datos de manera más rápida y eficiente a través del trait Database (por ejemplo; cambiando el tipo de la base de datos de una que funciona con la memoria RAM a una que funciona con un servidor externo que trabaja con consultas SQL).

Modules

class

Contiene las funciones para poder operar con los distintos tipos de bases de datos y funciones creadas para más simplicidad en caso de querer cambiar el tipo de la base de datos; implementando el patrón de diseño Abstract Factory para lograrlo junto con la interfaz Database.

db

Contiene las operaciones que deberán tener los distintos tipos de bases de datos del aplicativo; logrando así el patrón de diseño Abstract Factory; forzando a que los distintos tipos de bases de datos se implementen correctamente. /// Las bases de datos contienen los datos de los clientes y sus saldos.

error

Contiene una lista con todos los errores posibles de la aplicación para un manejo más eficiente de los mismos. También posee una función para procesar los mismos de una manera simple.