Module cribbage::domain

Cribbage Rules

Number of Players

Two or three people can play. Or four people can play two against two as partners. But Cribbage is basically best played by two people, and the rules that follow are for that number.

The Pack

The standard 52-card pack is used.

Rank of Cards: K (high), Q, J, 10, 9, 8, 7, 6, 5, 4, 3, 2, A.

The Draw, Shuffle and Cut

From a shuffled pack face down, each player cuts a card, leaving at least four cards at either end of the pack.

If both players cut cards of the same rank, each draws again. The player with the lower card deals the first hand. Thereafter, the turn to deal alternates between the two players, except that the loser of the game deals first if another game is played. The dealer has the right to shuffle last, and he presents the cards to the non-dealer for the cut prior to the deal. (In some games, there is no cut at this time.)

The Deal

The dealer distributes six cards face down to his opponent and himself, beginning with the opponent.

Object of the Game

The goal is to be the first player to score 121 points. (Some games are to 61 points.) Players earn points during play and for making various card combinations.

The Crib

Each player looks at his six cards and “lays away” (discards) two of them face down to reduce the hand to four. The four cards laid away together constitute “the crib”. The crib belongs to the dealer, but these cards are not exposed or used until after the hands have been played.

Before the Play

After the crib is laid away, the non-dealer cuts the pack. The dealer turns up the top card of the lower packet and places it face up on top of the pack. This card is the “starter.” If the starter is a jack, it is called “His Heels,” and the dealer pegs (scores) 2 points at once. The starter is not used in the play phase of Cribbage , but is used later for making various card combinations that score points.

The Play

After the starter is turned, the non-dealer lays one of his cards face up on the table. The dealer similarly exposes a card, then non-dealer again, and so on - the hands are exposed card by card, alternately except for a “Go,” (Pass) as noted below. Each player keeps his cards separate from those of his opponent.

As each person plays, he announces a running total of pips reached by the addition of the last card to all those previously played. (Example: The non-dealer begins with a four, saying “Four.” The dealer plays a nine, saying “Thirteen”.) The kings, queens and jacks count 10 each; every other card counts its pip value (the ace counts one).

The Go

During play, the running total of cards may never be carried beyond 31. If a player cannot add another card without exceeding 31, he or she says “Go” and the opponent pegs 1. After gaining the Go, the opponent must first lay down any additional cards he can without exceeding 31. Besides the point for Go, he may then score any additional points that can be made through pairs and runs (described later). If a player reaches exactly 31, he pegs two instead of one for Go.

The player who called Go leads for the next series of plays, with the count starting at zero. The lead may not be combined with any cards previously played to form a scoring combination; the Go has interrupted the sequence.

The person who plays the last card pegs one for Go, plus one extra if the card brings the count to exactly 31. The dealer is sure to peg at least one point in every hand, for he will have a Go on the last card if not earlier.

Pegging

The object in play is to score points by pegging. In addition to a Go, a player may score for the following combinations:

Fifteen: For adding a card that makes the total 15 Peg 2

Pair: For adding a card of the same rank as the card just played Peg 2 (Note that face cards pair only by actual rank: jack with jack, but not jack with queen.)

Triplet: For adding the third card of the same rank. Peg 6

Four: (also called “Double Pair” or “Double Pair Royal”) For adding the fourth card of the same rank Peg 12

Run (Sequence): For adding a card that forms, with those just played:

For a sequence of three Peg 3

For a sequence of four. Peg 4

For a sequence of five. Peg 5

(Peg one point more for each extra card of a sequence. Note that runs are independent of suits, but go strictly by rank; to illustrate: 9, 10, J, or J, 9, 10 is a run but 9, 10, Q is not)

It is important to keep track of the order in which cards are played to determine whether what looks like a sequence or a run has been interrupted by a “foreign card.” Example: Cards are played in this order: 8, 7, 7, 6. The dealer pegs 2 for 15, and the opponent pegs 2 for pair, but the dealer cannot peg for run because of the extra seven (foreign card) that has been played. Example: Cards are played in this order: 9, 6, 8, 7. The dealer pegs 2 for fifteen when he plays the six and pegs 4 for run when he plays the seven (the 6, 7, 8, 9 sequence). The cards were not played in sequential order, but they form a true run with no foreign card.

Counting the Hands

When play ends, the three hands are counted in order: non-dealer’s hand (first), dealer’s hand (second), and then the crib (third). This order is important because, toward the end of a game, the non-dealer may “count out” and win before the dealer has a chance to count, even though the dealer’s total would have exceeded that of the opponent. The starter is considered to be a part of each hand, so that all hands in counting comprise five cards. The basic scoring formations are as follows:

Combinations counts

Fifteen. Each combination of cards that totals 15 2

Pair. Each pair of cards of the same rank 2

Run. Each combination of three or more 1 cards in sequence (for each card in the sequence)

Flush.

Four cards of the same suit in hand 4 (excluding the crib, and the starter)

Four cards in hand or crib of the same 5 suit as the starter. (There is no count for four-flush in the crib that is not of same suit as the starter)

His Nobs. Jack of the same suit as starter in hand or crib 1

Combinations

In the above table, the word combination is used in the strict technical sense. Each and every combination of two cards that make a pair, of two or more cards that make 15, or of three or more cards that make a run, count separately.

Example: A hand (including the starter) comprised of 8, 7, 7, 6, 2 scores 8 points for four combinations that total 15: the 8 with one 7, and the 8 with the other 7; the 6, 2 with each of the two 7s. The same hand also scores 2 for a pair, and 6 for two runs of three (8, 7, 6 using each of the two 7s). The total score is 16. An experienced player computes the hand thus: “Fifteen 2, fifteen 4, fifteen 6, fifteen 8, and 8 for double run is 16.”

Note that the ace is always low and cannot form a sequence with a king. Further, a flush cannot happen during the play of the cards; it occurs only when the hands and the crib are counted.

Certain basic formulations should be learned to facilitate counting. For pairs and runs alone:

A. A triplet counts 6. A. Four of a kind counts 12. A. A run of three, with one card duplicated (double run) counts 8. A. A run of four, with one card duplicated, counts 10. A. A run of three, with one card triplicated (triple run), counts 15. A. A run of three, with two different cards duplicated, counts 16.

A PERFECT 29!

The highest possible score for combinations in a single Cribbage deal is 29, and it may occur only once in a Cribbage fan’s lifetime -in fact, experts say that a 29 is probably as rare as a hole-in-one in golf. To make this amazing score, a player must have a five as the starter (upcard) and the other three fives plus the jack of the same suit as the starter - His Nobs: 1 point - in his hand. The double pair royal (four 5s) peg another 12 points; the various fives used to hit 15 can be done four ways for 8 points; and the jack plus a 5 to hit 15 can also be done four ways for 8 points. Total = 29 points.

Miscellaneous

The following list includes many of the hands that may give the beginner some difficulty in counting. Note that no hand can make a count of 19, 25, 26, or 27. (In the chart below J stands for His Nobs, the jack of the same suit as the starter.

Muggins (optional) - not implemented.

Each player must count his hand (and crib) aloud and announce the total. If he overlooks any score, the opponent may say “Muggins” and then score the overlooked points for himself. For experienced players, the Muggins rule is always in effect and adds even more suspense to the game.

Game

Game may be fixed at either 121 points or 61 points. The play ends the moment either player reaches the agreed total, whether by pegging or counting one’s hand. If the non-dealer “goes out” by the count of his hand, the game immediately ends and the dealer may not score either his hand or the crib.

If a player wins the game before the loser has passed the halfway mark (did not reach 31 in a game of 61, or 61 in a game of 121), the loser is “lurched,” and the winner scores two games instead of one. A popular variation of games played to 121, is a “skunk” (double game) for the winner if the losing player fails to pass the three-quarter mark - 91 points or more - and it is a “double skunk” (quadruple game) if the loser fails to pass the halfway mark (61 or more points).

The Cribbage Board

The Cribbage board (see illustration) has four rows of 30 holes each, divided into two pairs of rows by a central panel. There are usually four (or two) additional holes near one end, called “game holes.” With the board come four pegs, usually in two contrasting colors. Note: There are also continuous track Cribbage boards available which, as the name implies, have one continuous line of 121 holes for each player.

The board is placed to one side between the two players, and each player takes two pegs of the same color. (The pegs are placed in the game holes until the game begins.) Each time a player scores, he advances a peg along a row on his side of the board, counting one hole per point. Two pegs are used, and the rearmost peg jumps over the first peg to show the first increment in score. After another increase in score, the peg behind jumps over the peg in front to the appropriate hole to show the player’s new score, and so on (see diagram next page). The custom is to “go down” (away from the game holes) on the outer rows and “come up” on the inner rows. A game of 61 is “once around” and a game of 121 is “twice around.” As noted previously, continuous line Cribbage boards are available.

If a Cribbage board is not available, each player may use a piece of paper or cardboard, marked thus:

Units 1, 2, 3, 4, 5, 6, 7, 8, 9, 10

Tens 10, 20, 30, 40, 50, 60

Two small markers, such as small coins or buttons, can substitute for pegs for counting in each row.

Strategy

The Crib.

If the dealer is discarding for the crib, he should “salt” it with the best possible cards, but at the same time retain good cards in his hand that can be used for high scoring. Conversely, for the non-dealer, it is best to lay out cards that will be the least advantageous for the dealer. Laying out a five would be the worst choice, for the dealer could use it to make 15 with any one of the ten-cards (10, J, Q, K). Laying out a pair is usually a poor choice too, and the same goes for sequential cards, such as putting both a six and seven in the crib. The ace and king tend to be good cards to put in the crib because it is harder to use them in a run.

The Play

As expected, the five makes for the worst lead in that there are so many ten-cards that the opponent can use to make a 15. Leading from a pair is a good idea, for even if the opponent makes a pair, the leader can play the other matching card from his hand and collect for a pair royal. Leading an ace or deuce is not a good idea, for these cards should be saved until later to help make a 15, a Go, or a 31. The safest lead is a four because this card cannot be used to make a 15 at the opponent’s very next turn. Finally, when the opponent leads a card that can either be paired or make 15, the latter choice is preferred.

During the play, it is advisable not to try to make a count of 21, for the opponent can then play one of the many 10-cards and make 31 to gain two points.

Modules

prelude