api
Module helper

Macros

impl_column_type_functions 🔒

This macro is used to generate datatype functions with lower style for ColumnDataTypeWrapper.

impl_column_type_functions_with_snake 🔒

This macro is used to generate datatype functions with snake style for ColumnDataTypeWrapper.

Structs

ColumnDataTypeWrapper

ColumnDataTypeWrapper is a wrapper of ColumnDataType and ColumnDataTypeExtension. It could be used to convert with ConcreteDataType.

Functions

convert_month_day_nano_to_pb

Converts an interval to google protobuf type [IntervalMonthDayNano].

convert_to_pb_decimal128

Convert common decimal128 to grpc decimal128 without precision and scale.

ddl_request_type 🔒

Returns the type name of the DdlRequest.

is_column_type_value_eq

Returns true if the pb type value is valid.

is_semantic_type_eq

Returns true if the pb semantic type is valid.

pb_value_to_value_ref

pb_values_to_values

pb_values_to_vector_ref

proto_value_type

Returns the ColumnDataTypeWrapper of the value.

push_vals

query_request_type 🔒

Returns the type name of the QueryRequest.

request_type

Returns the type name of the Request.

to_proto_value

Convert value into proto’s value.

value_to_grpc_value

values_with_capacity

vectors_to_rows