Function api::repos::list_user_repos

pub async fn list_user_repos() -> Result<Page<Repository>>

list_user_repos prints the names of repositories for the authenticated user.

Prerequisites

Create a personal access token in your GitHub account settings.

Store the access token as an environment variable, for example GITHUB_ACCESS_TOKEN.

In your Rust application, retrieve the access token from the environment variable using the dotenv crate.

Use the octocrab crate to make API requests to the GitHub API, passing the access token as an argument when creating a client.

Use the client to fetch information about repositories, such as a list of repositories for a specific user.

See also