gixor
Struct Gixor

pub struct Gixor { /* private fields */ }

Represents the main structure of Gixor.

Implementations

impl Gixor

pub fn load_or_default() -> Self

load the configuration file from the location. The default configuration is see Gixor::default.

pub fn load<P: AsRef<Path>>(path: P) -> Result<Self>

Parse the configuration file from the given path.

pub fn len(&self) -> usize

Returns the length of the repositories in the configuration.

pub fn is_empty(&self) -> bool

Returns true if the repositories in the configuration is empty.

pub fn base_path(&self) -> &Path

Returns the base path of this configuration.

pub fn store(&self) -> Result<()>

Store the configuration to the configuration path.

pub fn iter(&self) -> impl Iterator<Item = Boilerplate<'_>>

Iterate the boilerplate paths in the configuration.

pub fn find(&self, name: Name) -> Result<Vec<Boilerplate<'_>>>

Find the boilerplate by the name.

pub fn update_all(&self) -> Result<()>

Update the all of boilerplate repositories.

pub fn clone_all(&self) -> Result<()>

Clone all of the repositories in the configuration.