api
Module v1

Modules

add_column_location

Nested message and enum types in AddColumnLocation.

alter_database_expr

Nested message and enum types in AlterDatabaseExpr.

alter_table_expr

Nested message and enum types in AlterTableExpr.

auth_header

Nested message and enum types in AuthHeader.

column

Nested message and enum types in Column.

column_data_type_extension

Nested message and enum types in ColumnDataTypeExtension.

column_def

ddl_request

Nested message and enum types in DdlRequest.

flow

greptime_database_client

Generated client implementations.

greptime_database_server

Generated server implementations.

greptime_request

Nested message and enum types in GreptimeRequest.

greptime_response

Nested message and enum types in GreptimeResponse.

health_check_client

Generated client implementations.

health_check_server

Generated server implementations.

index

meta

prometheus_gateway_client

Generated client implementations.

prometheus_gateway_server

Generated server implementations.

promql_request

Nested message and enum types in PromqlRequest.

query_request

Nested message and enum types in QueryRequest.

region

value

Nested message and enum types in Value.

Structs

AddColumn

AddColumnLocation

AddColumns

AffectedRows

AlterDatabaseExpr

AlterTableExpr

AuthHeader

Basic

Column

ColumnDataTypeExtension

Type extension for some complex types

ColumnDef

ColumnOptions

Additional options for the column.

ColumnSchema

CreateDatabaseExpr

CreateFlowExpr

Create a flow to run the SQL when new data arrives.

CreateTableExpr

CreateViewExpr

Create a view

DdlRequest

“Data Definition Language” requests, that create, modify or delete the database structures but not the data. DdlRequest could carry more information than plain SQL, for example, the “table_id” in CreateTableExpr. So create a new DDL expr if you need it.

Decimal128

(hi: high 64 bits, lo: low 64 bits) are used to keep the decimal128 value.

DecimalTypeExtension

DeleteRequest

DeleteRequests

DropColumn

DropColumns

DropDatabaseExpr

DropFlowExpr

Drop a flow.

DropTableExpr

DropViewExpr

Drop a view

ExpireAfter

FlightMetadata

FlowId

GreptimeRequest

GreptimeResponse

HealthCheckRequest

HealthCheckResponse

InsertRequest

InsertRequests

IntervalMonthDayNano

Metrics

ModifyColumnType

ModifyColumnTypes

Mutation

Mutation contains updates to a set of rows.

Option

PromInstantQuery

PromRangeQuery

PromqlRequest

PromqlResponse

QueryContext

QueryRequest

RenameTable

RequestHeader

ResponseHeader

Row

RowDeleteRequest

RowDeleteRequests

RowInsertRequest

RowInsertRequests

Rows

SetColumnFulltext

SetDatabaseOptions

SetTableOptions

Status

TableId

TableName

Token

TruncateTableExpr

UnsetColumnFulltext

UnsetDatabaseOptions

UnsetTableOptions

Value

VectorTypeExtension

WalEntry

A WAL entry contains a list of mutations for a region to write.

Enums

Analyzer

ColumnDataType

JsonTypeExtension

OpType

Type of operation to rows.

SemanticType

Constants

GREPTIME_GRPC_DESC