api::v1
Module column_def

Constants

FULLTEXT_GRPC_KEY 🔒

Key used to store fulltext options in gRPC column options.

INVERTED_INDEX_GRPC_KEY 🔒

Key used to store inverted index options in gRPC column options.

SKIPPING_INDEX_GRPC_KEY 🔒

Key used to store skip index options in gRPC column options.

Functions

as_fulltext_option

Tries to construct a FulltextAnalyzer from the given analyzer.

contains_fulltext

Checks if the ColumnOptions contains fulltext options.

options_from_column_schema

Constructs a ColumnOptions from the given ColumnSchema.

options_from_fulltext

Tries to construct a ColumnOptions from the given FulltextOptions.

try_as_column_schema

Tries to construct a ColumnSchema from the given ColumnDef.