api::v1

Module helper

Functions

field_column_schema

Create a field ColumnSchema with column’s name and datatype. Other fields are left default. Useful when you just want to create a simple ColumnSchema without providing much struct fields.

row

Create a Row from ValueDatas. Useful when you don’t want to write much verbose codes.

tag_column_schema

Create a tag ColumnSchema with column’s name and datatype. Other fields are left default. Useful when you just want to create a simple ColumnSchema without providing much struct fields.

time_index_column_schema

Create a time index ColumnSchema with column’s name and datatype. Other fields are left default. Useful when you just want to create a simple ColumnSchema without providing much struct fields.