api::v1

Module index

Structs

BloomFilterLoc

BloomFilterLoc defines the location of a bloom filter.

BloomFilterMeta

BloomFilterMeta defines the metadata for a bloom filter.

InvertedIndexMeta

InvertedIndexMeta contains the metadata for a specific tag’s inverted index.

InvertedIndexMetas

InvertedIndexMetas defines the metadata for an inverted index within an inverted index blob.

InvertedIndexStats

InvertedIndexStats provides statistical data on a tag’s inverted index.

Enums

BitmapType

BitmapType defines the type of bitmap used for indexing.