api::v1

Module meta

Modules

cluster_client

Generated client implementations.

cluster_server

Generated server implementations.

ddl_task_request

Nested message and enum types in DdlTaskRequest.

heartbeat_client

Generated client implementations.

heartbeat_request

Nested message and enum types in HeartbeatRequest.

heartbeat_server

Generated server implementations.

lock_client

Generated client implementations.

lock_server

Generated server implementations.

mailbox_message

Nested message and enum types in MailboxMessage.

procedure_service_client

Generated client implementations.

procedure_service_server

Generated server implementations.

store_client

Generated client implementations.

store_server

Generated server implementations.

Structs

AlterDatabaseTask

Alter database tasks

AlterTableTask

AlterTableTasks

AskLeaderRequest

AskLeaderResponse

BatchDeleteRequest

BatchDeleteResponse

BatchGetRequest

BatchGetResponse

BatchPutRequest

BatchPutResponse

CompareAndPutRequest

CompareAndPutResponse

CreateDatabaseTask

CreateFlowTask

CreateTableTask

CreateTableTasks

CreateTriggerTask

CreateViewTask

Create a view task

DatanodeWorkloads

/ The workload types of the datanode.

DdlTaskRequest

DdlTaskResponse

DeleteRangeRequest

DeleteRangeResponse

DropDatabaseTask

DropFlowTask

DropTableTask

DropTableTasks

DropViewTask

Drop a view task

Error

FlowStat

FlownodeWorkloads

/ The workload types of the flownode.

FrontendWorkloads

/ The workload types of the frontend.

GrantedRegion

The granted region is the region that the meta server granted leases.

HeartbeatRequest

HeartbeatResponse

KeyValue

LockRequest

LockResponse

MailboxMessage

MetasrvNodeInfo

MetasrvPeersRequest

MetasrvPeersResponse

MigrateRegionRequest

MigrateRegionResponse

NodeInfo

Partition

PARTITION region_name VALUES LESS THAN (value_list)

Peer

PeerDict

ProcedureDetailRequest

ProcedureDetailResponse

ProcedureId

Procedure identifier

ProcedureMeta

ProcedureStateResponse

PutRequest

PutResponse

QueryProcedureRequest

RangeRequest

RangeResponse

Region

RegionLease

RegionRoute

RegionStat

RequestHeader

ResponseHeader

Table

TableRoute

TableRouteValue

This message is only for saving into store. TODO(weny): Remove it, Now, the upgrade tool is still dependent on it.

TimeInterval

TruncateTableTask

UnlockRequest

UnlockResponse

Enums

DdlTaskType

ErrorCode

ProcedureStatus

RegionRole

Role

Constants

PROTOCOL_VERSION