Module yksom::vm::function

Structs

Function
The VM’s underlying notion of a SOM function. SOM has two types of functions: methods and blocks. Roughly speaking, methods have names, may reference primitives, but can’t be nested; blocks are anonymous, can’t reference primitives, but can be nested. Although at a SOM level methods and blocks are separate types, they share nearly everything that makes them function-esque, so to avoid having two separate internal representations, both methods and blocks store Functions.