Struct yksom::vm::objects::ThinObj

#[repr(C)]
pub struct ThinObj { /* fields omitted */ }

A narrow pointer to #trait_id.

Implementations

impl ThinObj

pub fn new<U: Send>(obj: U) -> Gc<Self> where
*const U: CoerceUnsized<*const dyn Obj + 'static>,
U: Obj + 'static,

Create a new narrow pointer to U: #trait_id.

pub unsafe fn new_from_layout<U: Obj + 'static, F>(
layout: Layout,
init: F
) -> Gc<Self> where
F: FnOnce(*mut U),

Create a narrow pointer to U: #trait_id. layout must be at least big enough for an object of type U (but may optionally be bigger) and must have at least the same alignment that U requires (but may optionally have a bigger alignment). initwill be called with a pointer to uninitialised memory into which a fully initialised object of typeU*must* be written. Afterinitcompletes, the object will be considered fully initialised: failure to fully initialise it causes undefined behaviour. Note that if additional memory was requested beyond that needed to storeUthen that extra memory does not have to be initialised afterinit` completes.

pub fn as_gc(&self) -> Gc<dyn Obj>

pub unsafe fn recover_gc<T: Obj>(o: Gc<T>) -> Gc<ThinObj>

Convert a downcasted narrow trait object back into a normal narrow trait object. This will lead to undefined behaviour if o was not originally a narrow trait object.

pub fn downcast<U: Obj>(&self) -> Option<Gc<U>>

Try casting this narrow trait object to a concrete struct type U, returning Some(...) if this narrow trait object has stored an object of type U or None otherwise.