Trait yksom::vm::objects::Array

pub trait Array {
    fn at(self: Gc<Self>, vm: &VM, idx: usize) -> Result<Val, Box<VMError>>;

    unsafe fn unchecked_at(self: Gc<Self>, idx: usize) -> Val;

    fn at_put(
        self: Gc<Self>, 
        vm: &mut VM, 
        idx: usize, 
        val: Val
    ) -> Result<(), Box<VMError>>;

    fn iter(self: Gc<Self>) -> ArrayIterator;
}

Required methods

fn at(self: Gc<Self>, vm: &VM, idx: usize) -> Result<Val, Box<VMError>>

Return the item at index idx (using SOM indexing starting at 1) or an error if the index is invalid.

unsafe fn unchecked_at(self: Gc<Self>, idx: usize) -> Val

Return the item at index idx (using SOM indexing starting at 1). This will lead to undefined behaviour if the index is invalid.

fn at_put(
self: Gc<Self>,
vm: &mut VM,
idx: usize,
val: Val
) -> Result<(), Box<VMError>>

Set the item at index idx (using SOM indexing starting at 1) to val or return an error if the index is invalid.

fn iter(self: Gc<Self>) -> ArrayIterator

Iterate over this array's values.

Loading content...