Struct yksom::vm::somstack::SOMStack

pub struct SOMStack { /* fields omitted */ }

A contiguous stack of SOM values. This stack does minimal or no checking on important operations and users must ensure that they obey the constraints on each function herein, or undefined behaviour will occur.

The basic layout of the stack is as a series of function frames growing from the beginning of the stack upwards. On a 64-bit machine, a function frame looks roughly like:

0 : <arg 0> ... n * 8: <var 0> ... (n+m) * 8:

The compiler and VM treat <arg 0> as special: it always contains a reference to self (hence all functions have 1 extra parameter over those specified by the user). Functions are expected to be called with their arguments already in place on the stack, and the stack pointer pointing to after the arguments but before the variables (i.e. the function will then set up its variables however it wants). Similarly, when a function is returned, the return value is expected to be placed where <arg 0> was originally found (i.e. at the end of the previous function's working stack).

Implementations

impl SOMStack

pub fn new() -> Gc<Self>

pub fn is_empty(self: Gc<Self>) -> bool

Returns true if the stack contains no elements.

pub fn len(self: Gc<Self>) -> usize

Returns the number of elements in the stack.

pub fn remaining_capacity(self: Gc<Self>) -> usize

Returns the number of elements the stack can store before running out of room.

pub unsafe fn addr_of(self: Gc<Self>, n: usize) -> Gc<Val>

pub fn peek(self: Gc<Self>) -> Val

Returns the top-most value of the stack without removing it. If the stack is empty, calling this function will lead to undefined behaviour.

pub fn peek_at(self: Gc<Self>, off: usize) -> Val

Peeks at a value n items from the top of the stack.

pub fn peek_n(self: Gc<Self>, n: usize) -> Val

Peeks at a value n items from the top of the stack.

pub fn pop(mut self: Gc<Self>) -> Val

Pops the top-most value of the stack and returns it. If the stack is empty, calling this function will lead to undefined behaviour.

pub fn pop_n(mut self: Gc<Self>, n: usize) -> Val

Pops the top-most value of the stack and returns it. If the stack is empty, calling this function will lead to undefined behaviour.

pub fn push(mut self: Gc<Self>, v: Val)

Push v onto the end of the stack. You must previously have checked (using SOMStack::remaining_capacity) that there is room for this value: if there is not, undefined behaviour will occur.

pub fn set(self: Gc<Self>, n: usize, v: Val)

pub fn split_off(mut self: Gc<Self>, at: usize) -> Val

Splits the collection into two at the given index.

Returns a newly allocated NormalArray containing the elements in the range [at, len). After the call, the SOM stack will be left containing the elements [0, at) with its previous capacity unchanged.

pub fn truncate(mut self: Gc<Self>, len: usize)

Shortens the stack, keeping the first len elements and dropping the rest.