Crate gixor

Gixor is a tool to manage the boilerplate files (.gitignore). This is alternative tool of gibo written in Rust.

Also, this library provides an API of the Gitignore boilerplate management. The main structure of this library is Gixor.

Example of Dump the boilerplate

use gixor::{Gixor, Name, Result};

let gixor = Gixor::load("testdata/config.json").unwrap();
gixor.update_all().unwrap(); // clone or update all repositories, if needed.
let r = gixor::dump_boilerplates(&gixor, "-", // dump the boilerplate of rust, macos, linux, and windows into stdout.
    vec![Name::parse("rust"), Name::parse("macos"), Name::parse("linux"), Name::parse("windows")]);

Modules

alias

Structs

Boilerplate

Represents a boilerplate file.

Gixor

Represents the main structure of Gixor.

Name

The name of the boilerplate which contains the repository name and the boilerplate name. The repository name is Repository::name. The boilerplate name is the file stem of the boilerplate (gitignore) file.

Repository

Represents a repository of the boilerplates. The boilerplate repository is cloned into ${base_path}/${name}

Enums

GixorError

Represents an error of Gixor.

Traits

AliasManager

Provides the functions for management of the aliases.

RepositoryManager

Provides the functions for management of the boilerplate repositories.

Functions

dump_boilerplates

Write the the content of boilerplate corresponding the given names to the destination. If the destination is "-", the content is written to the stdout, and the dest is a directory, the content is written to the ${dest}/.gitignore. Otherwise, the content is written to the file of dest.

find_target_repositories

list_entries

Returns the boilerplate names listed in the given path (.gitignore).

Type Aliases

Result

Represents the result of Gixor.