api::prom_store

Module remote

Modules

chunk

Nested message and enum types in Chunk.

label_matcher

Nested message and enum types in LabelMatcher.

metric_metadata

Nested message and enum types in MetricMetadata.

read_request

Nested message and enum types in ReadRequest.

Structs

Chunk

Chunk represents a TSDB chunk. Time range [min, max] is inclusive.

ChunkedReadResponse

ChunkedReadResponse is a response when response_type equals STREAMED_XOR_CHUNKS. We strictly stream full series after series, optionally split by time. This means that a single frame can contain partition of the single series, but once a new series is started to be streamed it means that no more chunks will be sent for previous one. Series are returned sorted in the same way TSDB block are internally.

ChunkedSeries

ChunkedSeries represents single, encoded time series.

Exemplar

Label

LabelMatcher

Matcher specifies a rule, which can match or set of labels or not.

Labels

MetricMetadata

Query

QueryResult

ReadHints

ReadRequest

ReadRequest represents a remote read request.

ReadResponse

ReadResponse is a response when response_type equals SAMPLES.

Sample

TimeSeries

TimeSeries represents samples and labels for a single time series.

WriteRequest