api::v1

Module column_def

Constants

FULLTEXT_GRPC_KEY 🔒

Key used to store fulltext options in gRPC column options.

INVERTED_INDEX_GRPC_KEY 🔒

Key used to store inverted index options in gRPC column options.

SKIPPING_INDEX_GRPC_KEY 🔒

Key used to store skip index options in gRPC column options.

Functions

as_fulltext_option_analyzer

Tries to construct a FulltextAnalyzer from the given analyzer.

as_fulltext_option_backend

Tries to construct a FulltextBackend from the given backend.

as_skipping_index_type

Tries to construct a SkippingIndexType from the given skipping index type.

contains_fulltext

Checks if the ColumnOptions contains fulltext options.

contains_skipping

Checks if the ColumnOptions contains skipping index options.

options_from_column_schema

Constructs a ColumnOptions from the given ColumnSchema.

options_from_fulltext

Tries to construct a ColumnOptions from the given FulltextOptions.

options_from_inverted

Tries to construct a ColumnOptions for inverted index.

options_from_skipping

Tries to construct a ColumnOptions from the given SkippingIndexOptions.

try_as_column_schema

Tries to construct a ColumnSchema from the given ColumnDef.