api::v1

Module region

Modules

alter_request

Nested message and enum types in AlterRequest.

bulk_insert_request

Nested message and enum types in BulkInsertRequest.

compact_request

Nested message and enum types in CompactRequest.

region_client

Generated client implementations.

region_request

Nested message and enum types in RegionRequest.

region_server

Generated server implementations.

sync_request

Nested message and enum types in SyncRequest.

Structs

AddColumn

AddColumns

AlterRequest

AlterRequests

BulkInsertRequest

Request of bulk ingestion API.

CloseRequest

CompactRequest

CreateRequest

CreateRequests

Create a batch of regions at once, usually used to create multiple logical regions at once. Different engines can choose whether to support this request. Metric Engine needs it.

DeleteRequest

DeleteRequests

DropColumn

DropColumns

DropRequest

DropRequests

Same as CreateRequests, but for dropping regions.

FlushRequest

InsertRequest

InsertRequests

MetricManifestInfo

Manifest info for metric engine.

MitoManifestInfo

Manifest info for mito engine.

OpenRequest

QueryRequest

RegionColumnDef

The column definition of a region. Unlike the message ColumnDef in ddl.proto which is for clients outside GreptimeDB, this RegionColumnDef is for region requests use only. So it carries an extra field column_id that is known internally.

RegionRequest

RegionRequestHeader

RegionResponse

Regular

/ Regular compaction

StrictWindow

/ Strictly-windowed compaction.

SyncRequest

Sync request for region.

TruncateRequest