Struct yksom::vm::core::VM

pub struct VM {
    pub block_cls: Val,
    pub block2_cls: Val,
    pub block3_cls: Val,
    pub bool_cls: Val,
    pub cls_cls: Val,
    pub double_cls: Val,
    pub false_cls: Val,
    pub int_cls: Val,
    pub metacls_cls: Val,
    pub nil_cls: Val,
    pub obj_cls: Val,
    pub str_cls: Val,
    pub sym_cls: Val,
    pub system_cls: Val,
    pub true_cls: Val,
    pub false_: Val,
    pub nil: Val,
    pub system: Val,
    pub true_: Val,
    // some fields omitted
}

The core VM struct.

Fields
block_cls: Valblock2_cls: Valblock3_cls: Valbool_cls: Valcls_cls: Valdouble_cls: Valfalse_cls: Valint_cls: Valmetacls_cls: Valnil_cls: Valobj_cls: Valstr_cls: Valsym_cls: Valsystem_cls: Valtrue_cls: Valfalse_: Valnil: Valsystem: Valtrue_: Val
Methods

impl VM

pub fn new(classpath: Vec<String>) -> Self

pub fn compile(&mut self, path: &Path, inst_vars_allowed: bool) -> Val

Compile the file at path. inst_vars_allowed should be set to false only for those builtin classes which do not lead to run-time instances of Inst.

pub fn error(&self, error: &str) -> !

Inform the user of the error string error and then exit.

pub fn top_level_send(
&mut self,
rcv: Val,
msg: &str,
args: Vec<Val>
) -> Result<Val, Box<VMError>>

Send the message msg to the receiver rcv with arguments args.

pub fn frames_len(&self) -> usize

pub fn push_blockinfo(&mut self, blkinfo: BlockInfo) -> usize

Add blkinfo to the set of known BlockInfos and return its index.

pub fn set_blockinfo(&mut self, idx: usize, blkinfo: BlockInfo)

Update the BlockInfo at index idx to blkinfo.

pub fn new_inline_cache(&mut self) -> usize

Add an empty inline cache to the VM, returning its index.

pub fn inline_cache_lookup(
&mut self,
idx: usize,
rcv_cls: Val,
name: &str
) -> Result<Gc<Method>, Box<VMError>>

Lookup the method name in the class rcv_cls, utilising the inline cache at index idx.

pub fn instrs_len(&self) -> usize

How many instructions are currently present in the VM?

pub fn instrs_push(&mut self, instr: Instr, span: Span)

Push instr to the end of the current vector of instructions, associating span with it for the purposes of backtraces.

pub fn add_send(&mut self, send: (String, usize)) -> usize

Add the send send to the VM, returning its index. Note that sends are reused, so indexes are also reused.

pub fn add_string(&mut self, s: String) -> usize

Add the string s to the VM, returning its index. Note that strings are reused, so indexes are also reused.

pub fn add_symbol(&mut self, s: String) -> usize

Add the symbol s to the VM, returning its index. Note that symbols are reused, so indexes are also reused.

pub fn add_global(&mut self, s: String) -> usize

Add the global n to the VM, returning its index. Note that global names (like strings) are reused, so indexes are also reused.

pub fn get_global_or_nil(&self, name: &str) -> Val

Lookup the global name: if it has not been added, or has been added but not set, then self.nil will be returned.

pub fn get_legal_global(&self, i: usize) -> Result<Val, Box<VMError>>

Get the global at position i: if it has not been set (i.e. is ValKind::ILLEGAL) this will return Err(...).

pub fn set_global(&mut self, name: &str, v: Val)

Set the global name to the value v, overwriting the previous value (if any).