Trait yksom::vm::objects::Obj

pub trait Obj: Debug + GcLayout {
    fn dyn_objtype(&self) -> ObjType;

    fn get_class(&self, vm: &mut VM) -> Val;

    fn to_strval(&self, _: &mut VM) -> Result<Val, Box<VMError>> { ... }

    fn inst_var_lookup(&self, _: usize) -> Val { ... }

    fn inst_var_set(&self, _: usize, _: Val) { ... }

    fn add(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn and(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn div(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn double_div(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn modulus(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn mul(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn shl(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn sqrt(&self, _: &mut VM) -> Result<Val, Box<VMError>> { ... }

    fn sub(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn xor(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn ref_equals(&self, vm: &mut VM, other: Val) -> Result<Val, Box<VMError>> { ... }

    fn equals(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn not_equals(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn greater_than(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn greater_than_equals(
        &self, 
        _: &mut VM, 
        _: Val
    ) -> Result<Val, Box<VMError>> { ... }

    fn less_than(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }

    fn less_than_equals(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>> { ... }
}

The main SOM Object trait. Notice that code should almost never call these functions directly: you should instead call the equivalent function in the Val struct.

Required methods

fn dyn_objtype(&self) -> ObjType

What ObjType does this Val represent?

fn get_class(&self, vm: &mut VM) -> Val

What class is this object an instance of?

Loading content...
Provided methods

fn to_strval(&self, _: &mut VM) -> Result<Val, Box<VMError>>

Convert this object to a Val that represents a SOM string.

fn inst_var_lookup(&self, _: usize) -> Val

Lookup an instance variable in this object.

fn inst_var_set(&self, _: usize, _: Val)

Set an instance variable in this object.

fn add(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which adds other to this.

fn and(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which performs a bitwise and with other and this.

fn div(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which divides other from this.

fn double_div(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

fn modulus(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which performs a mod operation on this with other.

fn mul(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which multiplies other to this.

fn shl(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which shifts self other bits to the left.

fn sqrt(&self, _: &mut VM) -> Result<Val, Box<VMError>>

Produces a new Val which is the square root of this.

fn sub(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which subtracts other from this.

fn xor(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Produce a new Val which performs a bitwise xor with other and this

fn ref_equals(&self, vm: &mut VM, other: Val) -> Result<Val, Box<VMError>>

Is this Val reference equality equal to other? Only number types are likely to want to override this.

fn equals(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Does this Val equal other?

fn not_equals(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Does this Val not equal other?

fn greater_than(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Is this Val greater than other?

fn greater_than_equals(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Is this Val greater than or equal to other?

fn less_than(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Is this Val less than other?

fn less_than_equals(&self, _: &mut VM, _: Val) -> Result<Val, Box<VMError>>

Is this Val less than or equal to other?

Loading content...