Trait yksom::vm::val::NotUnboxable

pub trait NotUnboxable { }

Objects which impl this trait guarantee that they can only ever be stored boxed. Implementing this trait on objects which can be stored unboxed leads to undefined behaviour.